Struct basic_dsp_vector::combined_ops::Identifier

pub struct Identifier<T, N, D> where T: RealNumber, D: Domain, N: NumberSpace { /* fields omitted */ }

An identifier is just a placeholder for a data type used to ensure already at compile time that operations are valid.