Struct basic_dsp_vector::combined_ops::MultiOperation1

pub struct MultiOperation1<S, T, NO, DO> where S: ToSlice<T>, T: RealNumber, NO: NumberSpace, DO: Domain { /* fields omitted */ }

A multi operation which holds a vector and records all changes which need to be done to the vectors. By calling get on the struct all operations will be executed in one run.

Methods

impl<S, T, NO, DO> MultiOperation1<S, T, NO, DO> where S: ToSliceMut<T> + Owner, DspVec<S, T, NO, DO>: RededicateForceOps<DspVec<S, T, RealOrComplexData, TimeOrFrequencyData>>, T: RealNumber + 'static, NO: NumberSpace, DO: Domain

Holds two vectors and records all operations which shall be done on the vectors. A call to get then runs all recorded operations on the vectors and returns them. See the modules description for why this can be beneficial.

fn extend<NI2, DI2>(self,
vector: DspVec<S, T, NI2, DI2>)
-> MultiOperation2<S, T, NO, DO, NI2, DI2> where NI2: NumberSpace, DI2: Domain

Extends the operation to operate on one more vector.

fn get<B>(self,
buffer: &mut B)
-> Result<DspVec<S, T, NO, DO>, (ErrorReason, DspVec<S, T, NO, DO>)> where B: Buffer<S, T>

Executes all recorded operations on the stored vector.

fn add_ops<F, NT, DT>(self, operation: F) -> MultiOperation1<S, T, NT, DT> where F: Fn(Identifier<T, NO, DO>) -> Identifier<T, NT, DT>, DT: Domain, NT: NumberSpace

Adds new operations which will be executed with the next call to get

As a background: The function operation will be executed immediately. It only operated on Identifier types and these serve as placeholder for vectors. Every operation done to an Identifier is recorded and will be executed on vectors if get is called.

fn add_enum_op(&mut self, op: Operation<T>)

Allows to directly push an Operation enum to a MultiOperation1. This mainly exists as interop between Rust and other languages.