Struct basic_dsp_vector::combined_ops::PreparedOperation2

pub struct PreparedOperation2<T, NI1, DI1, NI2, DI2, NO1, DO1, NO2, DO2> where
    T: RealNumber,
    NI1: NumberSpace,
    DI1: Domain,
    NI2: NumberSpace,
    DI2: Domain,
    NO1: NumberSpace,
    DO1: Domain,
    NO2: NumberSpace,
    DO2: Domain,  { /* fields omitted */ }

An operation on two data vectors which has been prepared in advance.

The type arguments can be read like a function definition: (TI1, TI2) -> (TO1, TO2)

Methods

impl<T, NI1, DI1, NI2, DI2, NO1, DO1, NO2, DO2> PreparedOperation2<T, NI1, DI1, NI2, DI2, NO1, DO1, NO2, DO2> where
T: RealNumber + 'static,
NI1: NumberSpace,
DI1: Domain,
NI2: NumberSpace,
DI2: Domain,
NO1: NumberSpace,
DO1: Domain,
NO2: NumberSpace,
DO2: Domain,

An operation which can be prepared in advance and operates on one input and produces one output

fn add_ops<F, NT1, DT1, NT2, DT2>(
self,
operation: F
) -> PreparedOperation2<T, NI1, DI1, NI2, DI2, NT1, DT1, NT2, DT2> where
F: Fn(Identifier<T, NO1, DO1>, Identifier<T, NO2, DO2>) -> (Identifier<T, NT1, DT1>, Identifier<T, NT2, DT2>),
DT1: Domain,
NT1: NumberSpace,
DT2: Domain,
NT2: NumberSpace,

Adds new operations which will be executed with the next call to exec

As a background: The function operation will be executed immediately. It only operated on Identifier types and these serve as placeholder for vectors. Every operation done to an Identifier is recorded and will be executed on vectors if exec is called.

fn add_enum_op(&mut self, op: Operation<T>)

Allows to directly push an Operation enum to a PreparedOperation1. This mainly exists as interop between Rust and other languages.