Struct basic_dsp_vector::conv_types::ForeignComplexConvolutionFunction

pub struct ForeignComplexConvolutionFunction<T> where
    T: RealNumber,  {
    pub conv_function: extern fn(_: *const c_void, _: T) -> Complex<T>,
    pub conv_data: usize,
    pub is_symmetric: bool,
}

A complex function which can be constructed outside this crate.

Fields
conv_function: extern fn(_: *const c_void, _: T) -> Complex<T>

The function

conv_data: usize

The data which is passed to the window function

Actual data type is a const* c_void, but Rust doesn't allow that because it's unsafe so we store it as usize and transmute it when necessary. Callers should make very sure safety is guaranteed.

is_symmetric: bool

Indicates whether this function is symmetric around 0 or not. Symmetry is defined as self.calc(x) == self.calc(-x).

Methods

impl<T> ForeignComplexConvolutionFunction<T> where
T: RealNumber,

unsafe fn new(
function: extern fn(_: *const c_void, _: T) -> Complex<T>,
function_data: *const c_void,
is_symmetric: bool
) -> Self

Creates a new real function