Module basic_dsp_vector::numbers

Traits from the num crate which are used inside basic_dsp and extensions to those traits.

Structs

Complex
A complex number in Cartesian form.

Traits

DspNumber
A trait for a numeric value which at least supports a subset of the operations defined in this crate. Can be an integer or a floating point number. In order to have support for all operations in this crate a must implement the RealNumber.

Float

Num
The base trait for numeric types, covering 0 and 1 values, comparisons, basic numeric operations, and string conversion.

One
Defines a multiplicative identity element for Self.

RealNumber
A real floating pointer number intended to abstract over f32 and f64.

ToSimd
Associates a number type with a SIMD register type.

Zero
This trait is necessary so that we can define zero for types outside this crate. It calls the num_traits::Zero trait where possible.