Trait basic_dsp_vector::numbers::DspNumber

pub trait DspNumber: Num + Copy + Clone + Send + Sync + ToSimd + Debug + Signed + FromPrimitive + FFTnum + 'static { }

A trait for a numeric value which at least supports a subset of the operations defined in this crate. Can be an integer or a floating point number. In order to have support for all operations in this crate a must implement the RealNumber.