Struct basic_dsp_vector::DspVec

pub struct DspVec<S, T, N, D> where S: ToSlice<T>, T: RealNumber, N: NumberSpace, D: Domain {
    pub data: S,
    // some fields omitted
}

A 1xN (one times N elements) or Nx1 data vector as used for most digital signal processing (DSP) operations. All data vector operations consume the vector they operate on and return a new vector. A consumed vector must not be accessed again.

Vectors come in different flavors:

Time or Frequency domain

Real or Complex numbers

32bit or 64bit floating point numbers

The first two flavors define meta information about the vector and provide compile time information what operations are available with the given vector and how this will transform the vector. This makes sure that some invalid operations are already discovered at compile time. In case that this isn't desired or the information about the vector isn't known at compile time there are the generic DataVec32 and DataVec64 vectors available.

32bit and 64bit flavors trade performance and memory consumption against accuracy. 32bit vectors are roughly two times faster than 64bit vectors for most operations. But remember that you should benchmark first before you give away accuracy for performance unless however you are sure that 32bit accuracy is certainly good enough.

Fields
data: S

The underlying storage. self.len() should be called to find out how many elements in data contain valid data.

Methods

impl<S, T, N, D> DspVec<S, T, N, D> where S: ToSliceMut<T>, T: RealNumber, N: NumberSpace, D: Domain

fn convolve_mat(matrix: &[&Self],
impulse_response: &[&Self],
target: &mut S)
-> VoidResult

Convolves a vector of vectors (in this lib also considered a matrix) with a vector of impulse responses and stores the result in target.