Struct basic_dsp_vector::FixedLenBuffer

pub struct FixedLenBuffer<S, T> where
    S: ToSliceMut<T>,
    T: RealNumber,  { /* fields omitted */ }

A buffer which gets initalized with a data storage type and then always keeps that.

Methods

impl<S, T> FixedLenBuffer<S, T> where
S: ToSliceMut<T>,
T: RealNumber,

fn new(storage: S) -> FixedLenBuffer<S, T>

Creates a new buffer from a storage type. The buffer will internally hold its storage for it's complete life time.