Struct basic_dsp_vector::SingleBuffer

pub struct SingleBuffer<T> where
    T: RealNumber,  { /* fields omitted */ }

A buffer which stores a single vector and never shrinks.

Methods

impl<T> SingleBuffer<T> where
T: RealNumber,

fn new() -> SingleBuffer<T>

Creates a new buffer which is ready to be passed around.

fn with_capacity(len: usize) -> SingleBuffer<T>

Creates a new buffer which is ready to be passed around.