Trait basic_dsp_vector::ComplexIndexMut

pub trait ComplexIndexMut<Idx>: ComplexIndex<Idx> where Idx: Sized {
    fn complex_mut(&mut self, index: Idx) -> &mut Self::Output;
}

Like std::ops::IndexMut but with a different method name so that it can be used to implement a additional range accessor for complex data.

Note if indexers will return an empty array in case the vector isn't complex.

Required Methods

fn complex_mut(&mut self, index: Idx) -> &mut Self::Output

The method for complex indexing