Trait basic_dsp_vector::ConvolutionOps

pub trait ConvolutionOps<A, S, T, N, D> where
    S: ToSliceMut<T>,
    T: RealNumber,
    N: NumberSpace,
    D: Domain, {
    fn convolve_signal<B>(
        &mut self, 
        buffer: &mut B, 
        impulse_response: &A
    ) -> VoidResult
    where
        B: for<'a> Buffer<'a, S, T>;
}

Provides a convolution operation for types which at some point are slice based.

Required Methods

fn convolve_signal<B>(
&mut self,
buffer: &mut B,
impulse_response: &A
) -> VoidResult where
B: for<'a> Buffer<'a, S, T>,

Convolves self with the convolution function impulse_response. For performance it's recommended to use multiply both vectors in frequency domain instead of this operation.

Upcoming shanges

With version 0.5 this method might get renamed.

Failures

TransRes may report the following ErrorReason members:

VectorMustBeInTimeDomain: if self is in frequency domain.

VectorMetaDataMustAgree: in case self and impulse_response are not in the same number space and same domain.

InvalidArgumentLength: if self.points() < impulse_response.points().