Trait basic_dsp_vector::FloatIndex

pub trait FloatIndex<Idx> where
    Idx: Sized, {
    type Output: ?Sized;
    fn data(&self, index: Idx) -> &Self::Output;
}

Like std::ops::Index but with a different method name so that it can be used to implement an additional range accessor for float data.

Associated Types

type Output: ?Sized

Loading content...
Required methods

fn data(&self, index: Idx) -> &Self::Output

The method for float indexing. If the vector is real valued then this function just returns the values

Loading content...