Trait basic_dsp_vector::FrequencyMultiplication

pub trait FrequencyMultiplication<'a, S, T, C: 'a> where
    S: ToSliceMut<T>,
    T: RealNumber, {
    fn multiply_frequency_response(&mut self, frequency_response: C, ratio: T);
}

Provides a frequency response multiplication operations.

Required methods

fn multiply_frequency_response(&mut self, frequency_response: C, ratio: T)

Multiplies self with the frequency response function frequency_response.

In order to multiply a vector with another vector in frequency response use mul.

Assumptions

The operation assumes that the vector contains a full spectrum centered at 0 Hz. If half a spectrum or a FFT shifted spectrum is provided the operation will come back with invalid results.

Failures

TransRes may report the following ErrorReason members:

VectorMustBeComplex: if self is in real number space but frequency_response is in complex number space.

VectorMustBeInFreqDomain: if self is in time domain.

Loading content...