Trait basic_dsp_vector::GetMetaData

pub trait GetMetaData<T, N, D> where
    T: RealNumber,
    N: NumberSpace,
    D: Domain, {
    fn get_meta_data(&self) -> TypeMetaData<T, N, D>;
}

Gets the meta data of a type. This can be used to create a new type with the same meta data.

Example

use basic_dsp_vector::*;
let vector = vec!(1.0, 2.0, 3.0, 4.0, 5.0, 6.0).to_real_time_vec();
let meta_data = vector.get_meta_data();
let slice = &vector[0..2];
let slice = slice.to_dsp_vec(&meta_data);
assert_eq!(false, slice.is_complex());

Required methods

fn get_meta_data(&self) -> TypeMetaData<T, N, D>

Gets a copy of the vector meta data. This can be used to create new types with the same meta data.

Loading content...