Trait basic_dsp_vector::NumberSpace

pub trait NumberSpace: Debug + PartialEq + Clone {
    fn is_complex(&self) -> bool;
    fn to_complex(&mut self);
    fn to_real(&mut self);
}

Number space (real or complex) information.

Required Methods

fn is_complex(&self) -> bool

fn to_complex(&mut self)

For implementations which track meta data at runtime this method may be implemented to transition between different states. For all other implementations they may leave this empty.

fn to_real(&mut self)

See RealNumberSpace for some more details.