Trait basic_dsp_vector::PreciseStats

pub trait PreciseStats<T>: Sized {
    fn add_prec(&mut self, elem: T, index: usize, sumc: &mut T, rmsc: &mut T);
}

A trait for statistics which allows to add new values in a way so that the numerical uncertainty has less impact on the final results.

Required methods

fn add_prec(&mut self, elem: T, index: usize, sumc: &mut T, rmsc: &mut T)

Adds a new values to the statistics using the Kahan summation algorithm described here: https://en.wikipedia.org/wiki/Kahan_summation_algorithm

Loading content...