Trait basic_dsp_vector::RealInterpolationOps

pub trait RealInterpolationOps<S, T> where S: ToSliceMut<T>, T: RealNumber {
    fn interpolate_hermite<B>(&mut self,
                          buffer: &mut B,
                          interpolation_factor: T,
                          delay: T) where B: Buffer<S, T>;
    fn interpolate_lin<B>(&mut self,
                      buffer: &mut B,
                      interpolation_factor: T,
                      delay: T) where B: Buffer<S, T>;
}

Provides interpolation operations which are only applicable for real data vectors.

Failures

All operations in this trait fail with VectorMustBeReal if the vector isn't in the real number space.

Required Methods

fn interpolate_hermite<B>(&mut self,
buffer: &mut B,
interpolation_factor: T,
delay: T) where B: Buffer<S, T>

Piecewise cubic hermite interpolation between samples.

Unstable

Algorithm might need to be revised. This operation and interpolate_lin might be merged into one function with an additional argument in future.

fn interpolate_lin<B>(&mut self,
buffer: &mut B,
interpolation_factor: T,
delay: T) where B: Buffer<S, T>

Linear interpolation between samples.

Unstable

This operation and interpolate_hermite might be merged into one function with an additional argument in future.