Trait basic_dsp_vector::RededicateToOps

pub trait RededicateToOps<Other> where
    Other: MetaData, {
    fn rededicate(self) -> Other;
}

This trait allows to change a data type. The operations will convert a type to a different one and set self.len() to zero. However self.allocated_len() will remain unchanged. The use case for this is to allow to reuse the memory of a vector for different operations.

Required methods

fn rededicate(self) -> Other

Converts Self inot Other.

Loading content...