Trait basic_dsp_vector::Resize

pub trait Resize {
    fn resize(&mut self, len: usize);
}

A trait for storage types which are known to have the capability to increase their capacity.

Required methods

fn resize(&mut self, len: usize)

Resize a storage type. Must work for any value of len, however it's okay if after this method self.alloc_len() > len or in words: It's okay if the method allocates more memory than specified in the parameter.

Loading content...