Trait basic_dsp_vector::Stats

pub trait Stats<T>: Sized {
    fn empty() -> Self;

    fn empty_vec(len: usize) -> StatsVec<Self>;

    fn invalid() -> Self;

    fn merge(stats: &[Self]) -> Self;

    fn merge_cols(stats: &[StatsVec<Self>]) -> StatsVec<Self>;

    fn add(&mut self, elem: T, index: usize);
}

Operations on statistics.

Required Methods

fn empty() -> Self

Creates an empty statistics struct.

fn empty_vec(len: usize) -> StatsVec<Self>

Creates a vector of empty statistics structs.

fn invalid() -> Self

Creates a statistics struct which resembles an invalid result.

fn merge(stats: &[Self]) -> Self

Merges several statistics into one.

fn merge_cols(stats: &[StatsVec<Self>]) -> StatsVec<Self>

Merges several vectors of statistics into one vector.

fn add(&mut self, elem: T, index: usize)

Adds a new value to the statistics, all statistic fields get updated.