Trait basic_dsp_vector::SymmetricFrequencyToTimeDomainOperations

pub trait SymmetricFrequencyToTimeDomainOperations<S, T>: ToRealTimeResult where
    S: ToSliceMut<T>,
    T: RealNumber, {
    fn plain_sifft<B>(self, buffer: &mut B) -> TransRes<Self::RealTimeResult>
    where
        B: for<'a> Buffer<'a, S, T>;

    fn sifft<B>(self, buffer: &mut B) -> TransRes<Self::RealTimeResult>
    where
        B: for<'a> Buffer<'a, S, T>;

    fn windowed_sifft<B>(
        self, 
        buffer: &mut B, 
        window: &dyn WindowFunction<T>
    ) -> TransRes<Self::RealTimeResult>
    where
        B: for<'a> Buffer<'a, S, T>;
}

Defines all operations which are valid on DataVecs containing frequency domain data and the data is assumed to half of complex conjugate symmetric spectrum round 0 Hz where the 0 Hz element itself is real.

Failures

All operations in this trait set self.len() to 0 if the first element (0Hz) isn't real.

Required methods

fn plain_sifft<B>(self, buffer: &mut B) -> TransRes<Self::RealTimeResult> where
B: for<'a> Buffer<'a, S, T>,

Performs a Symmetric Inverse Fast Fourier Transformation under the assumption that self contains half of a symmetric spectrum starting from 0 Hz. This assumption isn't verified and no error is raised if the spectrum isn't symmetric. The reason for this is that there is no robust verification possible.

The argument indicates whether the resulting real vector should have 2*N or 2*N-1 points.

This version of the IFFT neither applies a window nor does it scale the vector.

fn sifft<B>(self, buffer: &mut B) -> TransRes<Self::RealTimeResult> where
B: for<'a> Buffer<'a, S, T>,

Performs a Symmetric Inverse Fast Fourier Transformation under the assumption that self contains half of a symmetric spectrum starting from 0 Hz. This assumption isn't verified and no error is raised if the spectrum isn't symmetric. The reason for this is that there is no robust verification possible.

The argument indicates whether the resulting real vector should have 2*N or 2*N-1 points.

fn windowed_sifft<B>(
self,
buffer: &mut B,
window: &dyn WindowFunction<T>
) -> TransRes<Self::RealTimeResult> where
B: for<'a> Buffer<'a, S, T>,

Performs a Symmetric Inverse Fast Fourier Transformation (SIFFT) and removes the FFT window. The SIFFT is performed under the assumption that self contains half of a symmetric spectrum starting from 0 Hz. This assumption isn't verified and no error is raised if the spectrum isn't symmetric. The reason for this is that there is no robust verification possible.

The argument indicates whether the resulting real vector should have 2*N or 2*N-1 points.

Loading content...