Trait basic_dsp_vector::ToSlice

pub trait ToSlice<T> {
    fn to_slice(&self) -> &[T];
    fn len(&self) -> usize;
    fn alloc_len(&self) -> usize;
    fn try_resize(&mut self, len: usize) -> VoidResult;
}

A trait to convert a type into a slice.

Required Methods

fn to_slice(&self) -> &[T]

Convert to a slice.

fn len(&self) -> usize

Length of a slice.

fn alloc_len(&self) -> usize

Gets the allocated length of a storage. It's expected that self.alloc_len() >= self.len() in all cases.

fn try_resize(&mut self, len: usize) -> VoidResult

Resizes the storage to support at least len elements or returns an error if resizing isn't supported.