Trait basic_dsp_vector::Vector

pub trait Vector<T>: MetaData + ResizeOps where T: RealNumber {
    fn delta(&self) -> T;
    fn set_delta(&mut self, delta: T);
    fn len(&self) -> usize;
    fn points(&self) -> usize;
    fn get_multicore_settings(&self) -> &MultiCoreSettings;
    fn set_multicore_settings(&mut self, settings: MultiCoreSettings);
    fn alloc_len(&self) -> usize;
}

A trait for vector types.

Required Methods

fn delta(&self) -> T

The x-axis delta. If domain is time domain then delta is in [s], in frequency domain delta is in [Hz].

fn set_delta(&mut self, delta: T)

Sets the x-axis delta. If domain is time domain then delta is in [s], in frequency domain delta is in [Hz].

fn len(&self) -> usize

The number of valid elements in the vector. This can be changed with the Resize trait.

fn points(&self) -> usize

The number of valid points. If the vector is complex then every valid point consists of two floating point numbers, while for real vectors every point only consists of one floating point number.

fn get_multicore_settings(&self) -> &MultiCoreSettings

Gets the multi core settings which determine how the work is split between several cores if the amount of data gets larger.

fn set_multicore_settings(&mut self, settings: MultiCoreSettings)

Sets the multi core settings which determine how the work is split between several cores if the amount of data gets larger.

fn alloc_len(&self) -> usize

Gets the number of allocated elements in the underlying vector. The allocated length may be larger than the length of valid points. In most cases you likely want to have lenor points instead.