Struct basic_dsp_vector::combined_ops::PreparedOperation1

pub struct PreparedOperation1<T, NI, DI, NO, DO> where T: RealNumber, NI: NumberSpace, DI: Domain, NO: NumberSpace, DO: Domain { /* fields omitted */ }

An operation on one data vector which has been prepared in advance.

The type arguments can be read like a function definition: TI1 -> TO1

Methods

impl<T, NI, DI, NO, DO> PreparedOperation1<T, NI, DI, NO, DO> where T: RealNumber + 'static, NI: NumberSpace, DI: Domain, NO: NumberSpace, DO: Domain

An operation which can be prepared in advance and operates on one input and produces one output

fn extend<NI2, DI2>(self,
number_space: NI2,
domain: DI2)
-> PreparedOperation2<T, NI, DI, NI2, DI2, NO, DO, NI2, DI2> where NI2: NumberSpace, DI2: Domain

Extends the operation to operate on one more vector.

fn add_ops<F, NT, DT>(self,
operation: F)
-> PreparedOperation1<T, NI, DI, NT, DT> where F: Fn(Identifier<T, NO, DO>) -> Identifier<T, NT, DT>, DT: Domain, NT: NumberSpace

Adds new operations which will be executed with the next call to exec

As a background: The function operation will be executed immediately. It only operated on Identifier types and these serve as placeholder for vectors. Every operation done to an Identifier is recorded and will be executed on vectors if exec is called.

fn add_enum_op(&mut self, op: Operation<T>)

Allows to directly push an Operation enum to a PreparedOperation1. This mainly exists as interop between Rust and other languages.