Trait basic_dsp_vector::ComplexToRealSetterOps

pub trait ComplexToRealSetterOps<A, T, N, D> where
    T: RealNumber,
    N: NumberSpace,
    D: Domain,
    A: GetMetaData<T, N, D>, {
    fn set_real_imag(&mut self, real: &A, imag: &A) -> VoidResult;

    fn set_mag_phase(&mut self, mag: &A, phase: &A) -> VoidResult;
}

Defines setters to create complex data from real data.

Failures

All operations in this trait set self.len() to 0 if the type isn't in the complex number space.

Required methods

fn set_real_imag(&mut self, real: &A, imag: &A) -> VoidResult

Overrides the self vectors data with the real and imaginary data in the given vectors. real and imag must have the same size.

fn set_mag_phase(&mut self, mag: &A, phase: &A) -> VoidResult

Overrides the self vectors data with the magnitude and phase data in the given vectors. Note that self vector will immediately convert the data into a real and imaginary representation of the complex numbers which is its default format. mag and phase must have the same size.

Loading content...