Trait basic_dsp_vector::CrossCorrelationArgumentOps

pub trait CrossCorrelationArgumentOps<S, T>: ToFreqResult where
    S: ToSliceMut<T>,
    T: RealNumber, {
    fn prepare_argument<B>(self, buffer: &mut B) -> Self::FreqResult
    where
        B: for<'a> Buffer<'a, S, T>;

    fn prepare_argument_padded<B>(self, buffer: &mut B) -> Self::FreqResult
    where
        B: for<'a> Buffer<'a, S, T>;
}

This trait allows to transform an argument so that it can be used for cross correlation. Refer to the description of CrossCorrelationOps for more details.

Required methods

fn prepare_argument<B>(self, buffer: &mut B) -> Self::FreqResult where
B: for<'a> Buffer<'a, S, T>,

Prepares an argument to be used for convolution. Preparing an argument includes two steps:

Calculate the plain FFT

Calculate the complex conjugate

fn prepare_argument_padded<B>(self, buffer: &mut B) -> Self::FreqResult where
B: for<'a> Buffer<'a, S, T>,

Prepares an argument to be used for convolution. The argument is zero padded to length of 2 * self.points() - 1 and then the same operations are performed as described for prepare_argument.

Loading content...