Trait basic_dsp_vector::RededicateForceOps

pub trait RededicateForceOps<Other> {
    fn rededicate_from_force(origin: Other) -> Self;
    fn rededicate_with_runtime_data(origin: Other,
                                is_complex: bool,
                                domain: DataDomain)
                                -> Self;
}

This trait allows to change a data type and performs the Conversion without any checks. RededicateOps provides the same functionality but performs runtime checks to avoid that data is interpreted the wrong way.

In almost all cases this trait shouldn't be used directly.

Required Methods

fn rededicate_from_force(origin: Other) -> Self

Make Other a Self without performing any checks.

fn rededicate_with_runtime_data(origin: Other,
is_complex: bool,
domain: DataDomain)
-> Self

Make Other a Self without performing any checks.

Try to set the domain and number space. There is no guarantee that this will succeed, since some rededication targets only support one domain and number space value. Failures will be silenty ignored (which is by design).