Trait basic_dsp_vector::ToComplexVector

pub trait ToComplexVector<S, T> where
    S: Sized + ToSlice<T>,
    T: RealNumber, {
    fn to_complex_time_vec(self) -> ComplexTimeVec<S, T>;

    fn to_complex_freq_vec(self) -> ComplexFreqVec<S, T>;
}

Conversion from a generic data type into a dsp vector with complex data.

Required methods

fn to_complex_time_vec(self) -> ComplexTimeVec<S, T>

Create a new vector in complex number space and time domain. delta can be changed after construction with a call of set_delta.

For complex vectors with an odd length the resulting value will have a zero length.

fn to_complex_freq_vec(self) -> ComplexFreqVec<S, T>

Create a new vector in complex number space and frequency domain. delta can be changed after construction with a call of set_delta.

For complex vectors with an odd length the resulting value will have a zero length.

Loading content...